d/rules: Dump reftest differences into the log as base64
Some of the reftests are failing reliably on buildds, but passing
reliably on the corresponding porterbox. We should inspect the
differences to check that they're close enough.

Debian package builds don't have an equivalent of $AUTOPKGTEST_ARTIFACTS
(as requested in https://bugs.launchpad.net/launchpad/+bug/1845159) so
base64-encoding the results and dumping them into the text log is the
best we can do right now.